fix(android): 强化文件/IO 安全并适配 Android 12
- 为各模块的 Activity 增加 android:exported="true"，避免 Android 12 启动崩溃
- 统一使用 decodeToString，避免字符集转换问题
- 修复 Scroller 类型计算与方向判断；NumberPicker 增加长度校验保护
- 使用 long 计算时间（Time.java），避免毫秒溢出
- 全面补充文件删除/重命名失败的判断与日志/异常上报，改进临时/损坏文件清理（字幕、字体、弹幕、7z、缓存等）
- PathHelper 目录检查失败抛异常；FormatSTL 改为 try-with-resources 正确关闭流
- 测试用例断言更严格，提升稳定性
- 启用 release 混淆，补充 ProGuard 规则；移除重复 targetSdk 设置
- .gitignore 新增 SonarCloud 报告忽略
- 代码质量脚本简化必填项校验逻辑

+++ anime_component/src/main/java/com/xyoye/anime_component/ui/dialog/date_picker/Scroller.java
@@ -386,7 +386,7 @@ public class Scroller  {
         mMode = FLING_MODE;
         mFinished = false;
 
-        float velocity =  (float) Math.sqrt(velocityX * velocityX + velocityY * velocityY);
+        float velocity =  (float) Math.sqrt((double) velocityX * velocityX + (double) velocityY * velocityY);
      
         mVelocity = velocity;
         final double l = Math.log(START_TENSION * velocity / ALPHA);
@@ -477,7 +477,7 @@ public class Scroller  {
      */
     public void setFinalX(int newX) {
         mFinalX = newX;
-        mDeltaX = mFinalX - mStartX;
+        mDeltaX = (float) mFinalX - mStartX;
         mFinished = false;
     }
 
@@ -490,7 +490,7 @@ public class Scroller  {
      */
     public void setFinalY(int newY) {
         mFinalY = newY;
-        mDeltaY = mFinalY - mStartY;
+        mDeltaY = (float) mFinalY - mStartY;
         mFinished = false;
     }
 
@@ -498,7 +498,7 @@ public class Scroller  {
      * @hide
      */
     public boolean isScrollingInDirection(float xvel, float yvel) {
-        return !mFinished && Math.signum(xvel) == Math.signum(mFinalX - mStartX) &&
-                Math.signum(yvel) == Math.signum(mFinalY - mStartY);
+        return !mFinished && Math.signum(xvel) == Math.signum((float) mFinalX - mStartX) &&
+                Math.signum(yvel) == Math.signum((float) mFinalY - mStartY);
     }
 }
